TST: move multithreaded test to pv_unittest
